requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
python-multipart>=0.0.6
pydantic>=2.6.0
pydantic-settings>=2.2.0
aiosqlite>=0.19.0
